fix: change TINYINT to INTEGER for pg compatibility

+++ migrations/20250903054808-create-review.js
@@ -30,7 +30,7 @@ module.exports = {
         onDelete: 'RESTRICT'
       },
       rating: {
-        type: Sequelize.TINYINT
+        type: Sequelize.INTEGER
       },
       comment: {
         type: Sequelize.TEXT
